+++ ostap/fitting/roostats.py
@@ -13,24 +13,20 @@ __author__  = "Vanya BELYAEV Ivan.Belyaev@itep.ru"
 __date__    = "2023-01-17"
 __all__     = (
     ##
-    'interval_PL'   , ## Profile-likelihood confidence interval or upper/lower limit 
-    'interval_FC'   , ## Feldman-Cousins    confidence interval or upper/lower limit 
-    'interval_BC'   , ## Bayesian           confidence interval or upper/lowee limir
-    'interval_MCMC' , ## MCMC               confidence interval or upper/lower limit 
-    ##
-    'create_MC_WS'  , ## helper function to create (and fill) ModelConfig & RooWorkspace
+    'ProfileLikelihoodInterval', ## Profile-likelihood confidence interval or upper/lower limit 
+    'FeldmanCousinsInterval'   , ## Feldman-Cousins    confidence interval or upper/lower limit 
+    'BayesianInterval'         , ## Bayesian           confidence interval or upper/lowee limir
+    'MCMCInterval'             , ## MCMC               confidence interval or upper/lower limit 
     ##
     )
 # =============================================================================
 import ostap.fitting.roofit
-from   ostap.core.ostap_types import string_types 
-import ROOT
+from   ostap.core.ostap_types import string_types, integer_types
+import ROOT, abc 
 # =============================================================================
 from   ostap.logger.logger import getLogger
 if '__main__' ==  __name__ : logger = getLogger ( 'ostap.fitting.roostats' )
 else                       : logger = getLogger ( __name__                 )
-# =============================================================================
-
 # =============================================================================
 ## Create (and fill) ModelConfig & RooWorkspace
 #  @code
@@ -40,7 +36,7 @@ else                       : logger = getLogger ( __name__                 )
 #  @endcode
 #  @see RooStats::ModelConfig 
 #  @see RooWorkspace  
-def create_MC_WS  ( pdf , params , dataset = None , ws = None , globobs = () ) :
+def create_MC_WS  ( pdf , params , dataset = None , ws = None , globobs = () , **kwargs ) :
     """ Create (and fill) ModelConfig & RooWorkspace
     >>> pdf     = ...
     >>> globobs = [ ... ] 
@@ -52,266 +48,468 @@ def create_MC_WS  ( pdf , params , dataset = None , ws = None , globobs = () ) :
     elif isinstance ( params , string_types    ) : params = [ params ]
     
     parameters = [ pdf.parameter ( p , dataset ) for p in params ]
+
+    from ostap.utils.cidict import cidict
+    from ostap.core.core    import cidict_fun
     
-    if not ws : ws = ROOT.RooWorkspace         ( 'WS%s' % pdf.name  , 'workspace' )
-        
-    mc = ROOT.RooStats.ModelConfig ( 'MC%s' % pdf.name  , 'model-config for %s' % pdf  , ws )
+    kw_args = cidict( transform = cidict_fun , **kwargs )
+    
+    if not ws :
+        wsname  = kw_args.pop ( 'ws_name'  , 'WS_%s'            % pdf.name )
+        wstitle = kw_args.pop ( 'ws_title' , 'workspace for %s' % pdf.name )
+        ws      = ROOT.RooWorkspace         ( wsname  , wstitle )
 
-    mc.SetPdf                  ( pdf.pdf     )
-    mc.SetObservables          ( pdf.vars    )
+    mcname  = kw_args.pop ( 'mc_name'  , 'MC_%s'               % pdf.name )
+    mctitle = kw_args.pop ( 'mc_title' , 'model-config for %s' % pdf.name )
+    mc      = ROOT.RooStats.ModelConfig ( mcname , mctitle , ws )
 
-    poi = ROOT.RooArgSet       ( *parameters ) 
-    mc.SetParametersOfInterest (  poi        )
+    mc.SetPdf                  ( pdf.pdf         )
+    mc.SetObservables          ( pdf.observables )
+
+    poi = ROOT.RooArgSet       ( *parameters     ) 
+    mc.SetParametersOfInterest (  poi            )
     
     nuis = [ v for v in pdf.params ( dataset ) if not v in pdf.vars and not v in poi ]
     nuis = ROOT.RooArgSet      ( *nuis )    
     mc.SetNuisanceParameters   (  nuis )
 
-    pdf.aux_keep.append ( poi  )
-    pdf.aux_keep.append ( nuis )
+    pdf.aux_keep.append        ( poi  )
+    pdf.aux_keep.append        ( nuis )
 
     if globobs :
+        if isinstance ( globobs , ROOT.RooAbsReal ) : globobs = [ globobs ] 
         gobs = [ pdf.parameter  ( v , dataset ) for v in globobs ]
         gobs = ROOT.RooArgSet   ( *gobs )
         mc.SetGlobalObservables (  gobs )
         pdf.aux_keep.append     (  gobs )
+
+    if kw_args :
+        logger.warning ( 'create ModelConfig: Ignore keyword arguments: %s' % [ k for k in kw_args ] )
         
-    
     return mc, ws
 
-
-# =============================================================================
-## helper method to get the actual innterval/limits
-def get_result ( interval , limit , par = None , ws = None , mc = None ) :
-    """Helper methdo to get the actual innterval/limits"""
+# ================================================================================
+## Helper (abstract) base class for confidence intervals and limits 
+class CLInterval(object)  :
+    """Helper (abstract) base class for confidence intervals and limits"""
     
-    if   limit and 0 < limit :
-        high   = interval.UpperLimit ( par ) if par else interval.UpperLimit ()
-        result = high
-    elif limit and 0 > limit :
-        low    = interval.LowerLimit ( par ) if par else interval.LowerLimit ()
-        result = low 
-    else :        
-        low    = interval.LowerLimit ( par ) if par else interval.LowerLimit ()
-        high   = interval.UpperLimit ( par ) if par else interval.UpperLimit ()
-        result = low , high 
+    def __init__ ( self            ,
+                   pdf             ,   ## pdf
+                   params          ,   ## parameter(s) of interest
+                   ws      = None  ,   ## existing workspace
+                   dataset = None  ,   ## dataset
+                   **kwargs        ) : ## other arguments 
+
+        self.__new_ws = True if not ws else False
+        ## use function, later move it into base class 
+        self.__mc , self.__ws = create_MC_WS ( pdf     = pdf     ,
+                                               params  = params  ,
+                                               dataset = dataset ,
+                                               ws      = ws      , **kwargs )
         
-    if ws : del ws
-    if mc : del mc  
+        self.__dataset  =  dataset
+        pp              = pdf.parameter ( params  ) 
+        self.__par      = self.__ws.var ( pp.name ) ## parameter as it is in workspace 
+        self.__interval = None 
 
-    return result
+    # =========================================================================
+    ## Abstract method to create the interval calculator
+    @abc.abstractmethod 
+    def make_interval ( self , level , dataset = None ) :
+        """Abstract method to create the interval calculator"""
+        pass
 
+    # =========================================================================
+    ## get the confidence interal at certain confidence level
+    #  @code
+    #  interval = ...
+    #  lower, upper = interval.interval ( 0.90 ) 
+    #  @endcode 
+    def interval ( self , level , dataset = None ) :
+        """get the confidence interal at certain confidence level
+        >>> interval = ...
+        >>> lower, upper = interval.interval ( 0.90 ) 
+        """
+        assert 0 < level < 1 , 'Invalid confidence level!'
+        self.__interval = self.make_interval ( level , dataset )
+        self.__interval.SetConfidenceLevel   ( level )
 
-# =============================================================================
-## Profile Likelihood 
-# =============================================================================
+        par    = self.par
+        
+        low    = self.__interval.LowerLimit  ( par ) if par else self.__interval.LowerLimit ()
+        high   = self.__interval.UpperLimit  ( par ) if par else self.__interval.UpperLimit ()
+  
+        return low, high 
 
-# =============================================================================
-## Get a profile likelihood confidence interval or upper/lower limit  
+    # =========================================================================
+    ## get the upper/lower limit at certain confidence level
+    #  @code
+    #  interval =...
+    #  upper    = interval.limit ( 0.90 , +1 ) 
+    #  lower    = interval.limit ( 0.90 , -1 ) 
+    #  @endcode
+    def limit    ( self , level , limit , dataset = None ) :
+        """Get the upper/lower limit at certain confidence level
+        >>> interval =...
+        >>> upper    = interval.limit ( 0.90 , +1 ) 
+        >>> lower    = interval.limit ( 0.90 , -1 ) 
+        """
+        assert 0 < level < 1 , 'Invalid confidence level!'
+        self.__interval = self.make_interval ( level , dataset ) 
+        self.__interval.SetConfidenceLevel   ( 1.0 - 0.5 * ( 1 - level ) ) ## ATTENTION!
+
+        par = self.par 
+        if   0 < limit : return self.__interval.UpperLimit ( par ) if par else self.__interval.UpperLimit ()
+        elif 0 > limit : return self.__interval.LowerLimit ( par ) if par else self.__interval.LowerLimit ()
+                
+    # =========================================================================
+    ## Get the upper limit at certain confidence level
+    #  @code
+    #  interval =...
+    #  upper    = interval.upper_limit ( 0.90 ) 
+    #  @endcode 
+    def upper_limit ( self , level , dataset = None ) :
+        """Get the upper limit at certain confidence level
+        >>> interval =...
+        >>> upper    = interval.upper_limit ( 0.90 ) 
+        """
+        return self.limit ( level , limit = +1 , dataset = dataset )
+    
+    # =========================================================================
+    ## get the lower limit at certain confidence level
+    #  @code
+    #  interval =...
+    #  lower    = interval.lower_limit ( 0.90 ) 
+    #  @endcode 
+    def lower_limit ( self , level , dataset = None ) :
+        """Get the lower limit at certain confidence level
+        >>> interval =...
+        >>> lower    = interval.lower_limit ( 0.90 ) 
+        """        
+        return self.limit ( level , limit = -1 , dataset = dataset )
+
+    # =========================================================================
+    @property
+    def ws ( self )  :
+        """'ws' : RooStats workspace"""
+        return self.__ws
+    @property
+    def workspace ( self )  :
+        """'workspace' : RooStats workspace"""
+        return self.ws
+    @property
+    def mc ( self )  :
+        """'mc' : RooStats ModelConfig object"""
+        return self.__mc    
+    @property
+    def model_config ( self )  :
+        """'model_config' : RooStats ModelConfig object"""
+        return self.mc
+    
+    @property
+    def par ( self ) :
+        """'par' : parameter as it is stored in workspace"""
+        return self.__par 
+    @par.setter
+    def par ( self , value ) :
+        self.__par = value 
+    
+        
+# ================================================================================
+## Profile Likelihood confidence interval
 #  @code
-#  pdf        =  ...
-#  dataset    = ...
-#  low, high  = interval_PL ( pdf , 'N' , 0.95 , dataset )
-#  low_limit  = interval_PL ( pdf , 'N' , 0.95 , dataset , limit = -1 )
-#  high_limit = interval_PL ( pdf , 'N' , 0.95 , dataset , limit = +1 )
-#  @endcode 
+#  interval     = ProfileLikelihoodInterval ( .... )
+#  lower, upper = interval.interval ( 0.90 ) 
+#  upper        = interval.limit ( 0.90 , +1 ) 
+#  lower        = interval.limit ( 0.90 , -1 ) 
+#  upper        = interval.upper_limit ( 0.90 ) 
+#  lower        = interval.lower_limit ( 0.90 ) 
+#  @endcode
 #  @see RooStats::ProfileLikelihoodCalculator
-def interval_PL ( pdf , param , level , dataset ,
-                  globobs = () , constraints = () , ws = None , limit = None ) :
-    """Get a profile likelihood confidence interval
-    >>> pdf        = ...
-    >>> dataset    = ...
-    >>> low, high  = interval_PL ( pdf , 'N' , 0.95 , dataset )
-    >>> low_limit  = interval_PL ( pdf , 'N' , 0.95 , dataset , limit = -1 )
-    >>> high_limit = interval_PL ( pdf , 'N' , 0.95 , dataset , limit = +1 )
+class ProfileLikelihoodInterval(CLInterval) :
+    """Profile Likelihood confidence interval
+    >>> interval     = ProfileLikelihoodInterval ( .... )
+    >>> lower, upper = interval.interval ( 0.90 ) 
+    >>> upper        = interval.limit ( 0.90 , +1 ) 
+    >>> lower        = interval.limit ( 0.90 , -1 ) 
+    >>> upper        = interval.upper_limit ( 0.90 ) 
+    >>> lower        = interval.lower_limit ( 0.90 ) 
     - see `ROOT.RooStats.ProfileLikelihoodCalculator`
     """
-    assert 0 < level < 1 , 'Invalid confidence level %s' % level
-
-    ws_new  = True if not ws else False  
-    mc , ws = create_MC_WS  ( pdf      = pdf      ,
-                              params   = param    ,
-                              dataset  = dataset  ,
-                              globobs  = globobs  ,
-                              ws       = ws       )
     
-    plc = ROOT.RooStats.ProfileLikelihoodCalculator ( dataset , mc )
+    ## create the interval 
+    def make_interval ( self , level , dataset = None ) :
+    ## create the interval 
+
+        ds = dataset if dataset else self.__dataset 
+        assert ds ,           'Invalid dataset!'
+
+        self.__plc = ROOT.RooStats.ProfileLikelihoodCalculator ( ds , self.mc )
+        self.__plc.SetConfidenceLevel ( level )
         
-    plc.SetConfidenceLevel ( level )
+        return self.__plc.GetInterval()
 
-    interval = plc.GetInterval()
-
-    pp       = pdf.parameter  ( param   ) 
-    par      = ws.var         ( pp.name )
-    
-    return get_result ( interval , limit , par , mc , ws ) if ws_new else get_result (  interval , limit , par ) 
-                         
-# =============================================================================
-## Feldman-Cousins 
-# =============================================================================
-
-# =============================================================================
-## Get  Feldman-Cousins  confidence interval or upper/lower limits 
+    ## ===========================================================================
+    #  make a plot
+    #  @see RooStats::LikelihoodIntervalPlot
+    def plot ( self ) :
+        """Make a plot
+        - see `ROOT.RooStats.LikelihoodIntervalPlot`
+        """
+        if self.__interval :
+            return  ROOT.RooStats.LikelihoodIntervalPlot( self.__interval )
+        
+# ================================================================================
+## Feldman-Cousins confidence interval
 #  @code
-#  pdf       = ...
-#  dataset   = ...
-#  low, high = conf_interval_FC ( pdf , 'N' , 0.95 , dataset )
-#  lower     = conf_interval_FC ( pdf , 'N' , 0.95 , dataset , limit = -1 )
-#  upper     = conf_interval_FC ( pdf , 'N' , 0.95 , dataset , limit = +1 )
-#  @endcode 
+#  interval     = FeldmanCousinsInterval ( .... )
+#  lower, upper = interval.interval ( 0.90 ) 
+#  upper        = interval.limit ( 0.90 , +1 ) 
+#  lower        = interval.limit ( 0.90 , -1 ) 
+#  upper        = interval.upper_limit ( 0.90 ) 
+#  lower        = interval.lower_limit ( 0.90 ) 
+#  @endcode
 #  @see RooStats::FeldmanCousins
-def interval_FC ( pdf , param , level , dataset , 
-                  globobs           = ()    ,
-                  constraints       = ()    , 
-                  ws                = None  ,
-                  nbins             = 200   ,
-                  adaptive_sampling = True  ,
-                  fluctuate         = False , 
-                  limit             = None ) :
-    """Get Felddman-Cousins confidence interval
-    >>> pdf       = ...
-    >>> dataset   = ...
-    >>> low, high = conf_interval_FC ( pdf , 'N' , 0.95 , dataset )
+class FeldmanCousinsInterval(CLInterval) :
+    """Feldman-Cousins confidence interval
+    >>> interval     = FeldmanCousinsInterval ( .... )
+    >>> lower, upper = interval.interval ( 0.90 ) 
+    >>> upper        = interval.limit ( 0.90 , +1 ) 
+    >>> lower        = interval.limit ( 0.90 , -1 ) 
+    >>> upper        = interval.upper_limit ( 0.90 ) 
+    >>> lower        = interval.lower_limit ( 0.90 ) 
     - see `ROOT.RooStats.FeldmanCousins`
     """
-    assert 0 < level < 1 , 'Invalid confidence level %s' % level
-
-    ws_new  = True if not ws else False  
-    mc , ws = create_MC_WS  ( pdf      = pdf      ,
-                              params   = param    ,
-                              dataset  = dataset  ,
-                              globobs  = globobs  ,
-                              ws       = ws       )
     
-    fcc = ROOT.RooStats.FeldmanCousins ( dataset , mc )
-    
-    fcc.SetConfidenceLevel      ( level     )
-    fcc.FluctuateNumDataEntries ( fluctuate ) 
-    fcc.UseAdaptiveSampling     ( True if adaptive_sampling else False  )
-    fcc.SetNBins                ( nbins )
-    
-    interval = fcc.GetInterval()
+    def __init__ ( self              ,
+                   pdf               ,   ## pdf
+                   params            ,   ## parameter(s) of interest
+                   ws        = None  ,   ## existing workspace
+                   dataset   = None  ,   ## dataset
+                   fluctuate = False ,   ## for RooStats.FeldmanCousins.FluctuateNumDataEntries
+                   adaptive  = True  ,   ## for RooStats.FeldmanCousins.UseAdaptiveSampling 
+                   nbins     = 200   ,   ## for RooStats.FeldmanCousins.SetNbins                    
+                   **kwargs          ) : ## other arguments 
+        
 
-    pp       = pdf.parameter  ( param   ) 
-    par      = ws.var         ( pp.name )
-    
-    return get_result ( interval , limit , par , mc , ws ) if ws_new else get_result ( interval , limit , par ) 
+        CLInterval.__init__ ( self,
+                              pdf     = pdf     ,
+                              params  = params  ,
+                              ws      = ws      ,
+                              dataset = dataset , **kwargs )
+
+        assert isinstance ( nbins , integer_types ) and 10 < nbins ,'Inavlid number of bins!'
+        
+        self.__fluctuate = True if fluctuate else False
+        self.__adaptive  = True if adaptive  else False
+        self.__nbins     = nbins
+        
+    ## create the interval 
+    def make_interval ( self , level , dataset = None ) :
+        """Create the interval"""
+        
+        ds = dataset if dataset else self.__dataset 
+        assert ds ,           'Invalid dataset!'
+
+        self.__fc = ROOT.RooStats.FeldmanCousins( ds , self.mc )
+        self.__fc.SetConfidenceLevel ( level )
+
+        self.__fc.FluctuateNumDataEntries ( self.fluctuate ) 
+        self.__fc.UseAdaptiveSampling     ( self.adaptive  )
+        self.__fc.SetNBins                ( self.nbins     )
+        
+        return self.__fc.GetInterval()
+
+    @property
+    def fluctuate ( self ) :
+        """'fluctuate': parameter for `RooStats.FeldmanCousins.FluctuateNumDataEntries`"""
+        return self.__fluctuate 
+    @property
+    def adaptive  ( self ) :
+        """'adaptive': parameter for `RooStats.FeldmanCousins.UseAdaptiveSampling`"""
+        return self.__adaptive
+    @property
+    def nbins     ( self ) :
+        """'nbins': parameter for `RooStats.FeldmanCousins.SetNBins`"""
+        return self.__nbins
 
 
-# =============================================================================
-## Bayesian calcualtor 
-# =============================================================================
-
-# =============================================================================
-## Get  Bayesian confidence interval or upper/lower limit
+# ================================================================================
+## Bayesian confidence interval
 #  @code
-#  pdf       = ...
-#  dataset   = ...
-#  low, high = interval_BC ( pdf , 'N' , 0.95 , dataset )
-#  lower     = interval_BC ( pdf , 'N' , 0.95 , dataset , limit = -1 )
-#  upper     = interval_BC ( pdf , 'N' , 0.95 , dataset , limit = +1 )
-#  @endcode 
-#  @see RooStats::BayesianCalculator
-def interval_BC ( pdf , param , level , dataset , 
-                  globobs           = ()   ,
-                  constraints       = ()   ,
-                  ws                = None ,
-                  prior             = None , 
-                  limit             = None ) :
-    """Get Bayesian confidence interval
-    >>> pdf       = ...
-    >>> dataset   = ...
-    >>> low, high = conf_interval_BC ( pdf , 'N' , 0.95 , dataset )
-    - see `ROOT.RooStats.BayesianCalculator`
+#  interval     = BayesianInterval ( .... )
+#  lower, upper = interval.interval ( 0.90 ) 
+#  upper        = interval.limit ( 0.90 , +1 ) 
+#  lower        = interval.limit ( 0.90 , -1 ) 
+#  upper        = interval.upper_limit ( 0.90 ) 
+#  lower        = interval.lower_limit ( 0.90 ) 
+#  @endcode
+#  @see RooStats::BayesianCalcualtor
+class BayesianInterval(CLInterval) :
+    """Bayesian confidence interval
+    >>> interval     = BAyesianInterval ( .... )
+    >>> lower, upper = interval.interval ( 0.90 ) 
+    >>> upper        = interval.limit ( 0.90 , +1 ) 
+    >>> lower        = interval.limit ( 0.90 , -1 ) 
+    >>> upper        = interval.upper_limit ( 0.90 ) 
+    >>> lower        = interval.lower_limit ( 0.90 ) 
+    - see `ROOT.RooStats.FeldmanCousins`
     """
-    assert 0 < level < 1 , 'Invalid confidence level %s' % level
-
-    ws_new  = True if not ws else False  
-    mc , ws = create_MC_WS  ( pdf      = pdf      ,
-                              params   = param    ,
-                              dataset  = dataset  ,
-                              globobs  = globobs  ,
-                              ws       = ws       )
     
-    pp  = pdf.parameter  ( param )
-    
-    from   ostap.fitting.pdfbasic import APDF1 
-    if   prior and isinstance ( prior , APDF1  ) :
-        mc.SetPriorPdf( prior.pdf )
-    elif prior and isinstance ( prior , ROOT.RooAbsPdf ) :
-        mc.SetPriorPdf( prior  )
-    else :
-        ws.factory("Uniform::prior(%s)" % pp.name )
-        mc.SetPriorPdf(ws.pdf("prior"))
-    
-    bc = ROOT.RooStats.BayesianCalculator ( dataset , mc )
-    bc.SetConfidenceLevel  ( level )
-    
-    bcInt = bc.GetInterval()
+    def __init__ ( self              ,
+                   pdf               ,   ## pdf
+                   params            ,   ## parameter(s) of interest
+                   ws        = None  ,   ## existing workspace
+                   dataset   = None  ,   ## dataset
+                   prior     = None  ,   ## Bayesin prior 
+                   **kwargs          ) : ## other arguments 
+        
+        ## initialize the Base class
+        CLInterval.__init__ ( self,
+                              pdf     = pdf     ,
+                              params  = params  ,
+                              ws      = ws      ,
+                              dataset = dataset , **kwargs )
+        
+        from   ostap.fitting.pdfbasic import APDF1
+        
+        assert prior is None or \
+               isinstance ( prior , APDF1          ) or \
+               isinstance ( prior , ROOT.RooAbsPdf ) ,  \
+               "Invalid prior!"
 
-    interval = bc.GetInterval()
+        self.__prior = prior
+        
+        if   prior and isinstance ( prior , APDF1  ) :
+            self.mc.SetPriorPdf ( prior.pdf )
+        elif prior and isinstance ( prior , ROOT.RooAbsPdf ) :
+            self.mc.SetPriorPdf ( prior  )
+        else :
+            self.ws.factory("Uniform::prior(%s)" % self.par.name )
+            self.mc.SetPriorPdf ( self.ws.pdf("prior") )
 
-    par      = None 
-    return get_result ( interval , limit , par , mc , ws ) if ws_new else get_result ( interval , limit , par ) 
+        ## important to reset it here! 
+        self.par = None
 
+    ## create the interval 
+    def make_interval ( self , level , dataset = None ) :
+        """Create the interval"""
+        
+        ds = dataset if dataset else self.__dataset 
+        assert ds ,           'Invalid dataset!'
 
-# =============================================================================
-## MCMC Calculator
-# =============================================================================
+        self.__bc = ROOT.RooStats.BayesianCalculator ( ds , self.mc )
+        self.__bc.SetConfidenceLevel ( level )
+        
+        return self.__bc.GetInterval()
 
+    @property
+    def prior     ( self ) :
+        """'prior': prior for `RooStats.BayesianCalcualtor`"""
+        return self.__prior
 
-# =============================================================================
-## Get MCMC confidence interval
+    ## ===========================================================================
+    #  make a plot
+    #  @see RooStats::BayesianCalcialtor::GetPosteriorPlot
+    def plot ( self ) :
+        """Make a plot
+        - see `ROOT.RooStats.BayesianCalculator.GetPosteriorPlot`
+        """
+        if self.__bc : return self.__bc.GetPosteriorPlot() 
+
+# ================================================================================
+## MCMC confidence interval
 #  @code
-#  pdf       = ...
-#  dataset   = ...
-#  low, high = interval_MCMC ( pdf , 'N' , 0.95 , dataset )
-#  lower     = interval_MCMC ( pdf , 'N' , 0.95 , dataset , limit = -1 )
-#  upper     = interval_MCMC ( pdf , 'N' , 0.95 , dataset , limit = +1 )
-#  @endcode 
+#  interval     = MCMCInterval ( .... )
+#  lower, upper = interval.interval ( 0.90 ) 
+#  upper        = interval.limit ( 0.90 , +1 ) 
+#  lower        = interval.limit ( 0.90 , -1 ) 
+#  upper        = interval.upper_limit ( 0.90 ) 
+#  lower        = interval.lower_limit ( 0.90 ) 
+#  @endcode
 #  @see RooStats::MCMCCalculator
-def interval_MCMC ( pdf , param , level , dataset , 
-                    globobs      = ()     ,
-                    constraints  = ()     ,
-                    ws           = None   ,
-                    nbins        = 200    ,
-                    burnsteps    = 500    ,
-                    numiters     = 100000 ,
-                    leftfraction = 0.5    ,
-                    limit        = None   ) :    
-    """Get MCMC confidence interval
-    >>> pdf       = ...
-    >>> dataset   = ...
-    >>> low, high = interval_FC   ( pdf , 'N' , 0.95 , dataset )
-    >>> lower     = interval_MCMC ( pdf , 'N' , 0.95 , dataset , limit = -1 )
-    >>> upper     = interval_MCMC ( pdf , 'N' , 0.95 , dataset , limit = +1 )
+class MCMCInterval(CLInterval) :
+    """MCMC confidence interval
+    >>> interval     = MCMCInterval  ( .... )
+    >>> lower, upper = interval.interval ( 0.90 ) 
+    >>> upper        = interval.limit ( 0.90 , +1 ) 
+    >>> lower        = interval.limit ( 0.90 , -1 ) 
+    >>> upper        = interval.upper_limit ( 0.90 ) 
+    >>> lower        = interval.lower_limit ( 0.90 ) 
     - see `ROOT.RooStats.MCMCCalculator`
     """
-    assert 0 < level < 1 , 'Invalid confidence level %s' % level
-
-    ws_new  = True if not ws else False  
-    mc , ws = create_MC_WS  ( pdf      = pdf      ,
-                              params   = param    ,
-                              dataset  = dataset  ,
-                              globobs  = globobs  ,
-                              ws       = ws       )
     
-    mcmc= ROOT.RooStats.MCMCCalculator ( dataset , mc )
+    def __init__ ( self                  ,
+                   pdf                   ,    ## pdf
+                   params                ,   ## parameter(s) of interest
+                   ws           = None   ,   ## existing workspace
+                   dataset      = None   ,   ## dataset
+                   burnsteps    = 500    ,   ## for RooStats.MCMCCalculator.SetNumBurnInSteps 
+                   iterations   = 100000 ,   ## for RooStats.MCMCCalculator.SetNumIters  
+                   leftfraction = 0.5    ,   ## for RooStats.MCMCCalculator.SetNumIters  
+                   nbins        = 200    ,   ## for RooStats.MCMCCalculator.SetLeftSideTailFraction
+                   **kwargs              ) : ## other arguments 
+        
+        CLInterval.__init__ ( self,
+                              pdf     = pdf     ,
+                              params  = params  ,
+                              ws      = ws      ,
+                              dataset = dataset , **kwargs )
 
-    mcmc.SetConfidenceLevel       ( level        )
-    
-    mcmc.SetNumBins               ( nbins        )
-    mcmc.SetNumBurnInSteps        ( burnsteps    )
-    mcmc.SetNumIters              ( numiters     )
-    mcmc.SetLeftSideTailFraction  ( leftfraction )
-    
-    interval = mcmc.GetInterval()
+        assert isinstance ( nbins      , integer_types ) and  10 < nbins      ,'Inavlid number of bins!'
+        assert isinstance ( burnsteps  , integer_types ) and  10 < burnsteps  , "Invalid nurnsteps!"
+        assert isinstance ( iterations , integer_types ) and 100 < iterations , "Invalid nurnsteps!"
+        assert 0<= leftfraction <1                                            , 'Invalid leftfraction!'
 
-    pp       = pdf.parameter  ( param   ) 
-    par      = ws.var         ( pp.name )
-    
-    return get_result ( interval , limit , par , mc , ws ) if ws_new else get_result ( interval , limit , par ) 
+        self.__nbins        = nbins
+        self.__burnsteps    = burnsteps
+        self.__iterations   = iterations
+        self.__leftfraction = leftfraction
+        
+    ## create the interval 
+    def make_interval ( self , level , dataset = None ) :
+        """Create the interval"""
+        
+        ds = dataset if dataset else self.__dataset 
+        assert ds ,           'Invalid dataset!'
 
+        self.__mcmc = ROOT.RooStats.MCMCCalculator ( ds , self.mc )
+        self.__mcmc.SetConfidenceLevel ( level )
+
+        self.__mcmc.SetNumBins              ( self.nbins        )
+        self.__mcmc.SetNumBurnInSteps       ( self.burnsteps    ) 
+        self.__mcmc.SetNumIters             ( self.iterations   )
+        self.__mcmc.SetLeftSideTailFraction ( self.leftfraction )
+
+        return self.__mcmc.GetInterval()
+
+    @property
+    def nbins        ( self ) :
+        """'nbins': parameter for `RooStats.MCMCCalculator.SetNumBins`"""
+        return self.__nbins
+    @property
+    def burnsteps    ( self ) :
+        """'burnsteps': parameter for `RooStats.MCMCCalculator.SetNumBurnInSteps`"""
+        return self.__burnsteps
+    @property
+    def iterations   ( self ) :
+        """'iterations': parameter for `RooStats.MCMCCalculator.SetNumIter`"""
+        return self.__iterations
+    @property
+    def leftfraction ( self ) :
+        """'leftfraction': parameter for `RooStats.MCMCCalculator.SetLeftSideTailFraction`"""
+        return self.__leftfraction
+
+    ## ===========================================================================
+    #  make a plot
+    #  @see RooStats::MCMCIntervalPlot
+    def plot ( self ) :
+        """Make a plot
+        - see `ROOT.RooStats.MCMCIntervallPlot`
+        """
+        if self.__interval :
+            return ROOT.RooStats.MCMCIntervalPlot(mcInt)( self.__interval )
+    
 
 # =============================================================================
 if '__main__' == __name__ :